Population and brain cancer cases in New Mexico counties during 1973--1991.

A dataset containing the population count and number of brain cancer cases in the counties of New Mexico during the years 1973--1991. The population numbers are interpolations from the censuses conducted in 1973, 1982, and 1991. Interpolations were done using a quadratic function of time. Thus the year-to-year changes are overly smooth but match the census numbers in the three years mentioned.

NM_popcas

Format

A data frame with 608 rows and 4 variables:

year

Integer; the year the cases were recorded.

county

Character; the name of the county (no spaces).

population

Integer; the population in that county and year.

count

Integer; the number of brain cancer cases in that county and year.