Creates a new

Takes a data.table containing the column stream and possibly other columns, and creates a new data.table with a column for streamset added to the columns in the supplied table, according to the argument streamsets. The key colums of the resulting data.table can be specified.

join_streamsets(table, streamsets, keys = c("streamsets"))

Arguments

table

A data.table with column stream and other columns (but none for streamset).

streamsets

A list of stream sets, elements being vectors of integers representing different data streams.

keys

Character vector of one or more column names; these columns are set as key columns in the output data.table.

Value

A new data.table with a column for streamset added to the supplied table.

Examples

tab <- as.data.table(expand.grid(location = 1:2,
                                 duration = 1:3,
                                 stream = 1:2))
scanstatistics:::join_streamsets(tab, list(1, 2, 1:2),
                                 keys = c("streamset", "duration"))

#>     streamset duration stream location
#>  1:         1        1      1        1
#>  2:         1        1      1        2
#>  3:         1        2      1        1
#>  4:         1        2      1        2
#>  5:         1        3      1        1
#>  6:         1        3      1        2
#>  7:         2        1      2        1
#>  8:         2        1      2        2
#>  9:         2        2      2        1
#> 10:         2        2      2        2
#> 11:         2        3      2        1
#> 12:         2        3      2        2
#> 13:         3        1      1        1
#> 14:         3        1      1        2
#> 15:         3        1      2        1
#> 16:         3        1      2        2
#> 17:         3        2      1        1
#> 18:         3        2      1        2
#> 19:         3        2      2        1
#> 20:         3        2      2        2
#> 21:         3        3      1        1
#> 22:         3        3      1        2
#> 23:         3        3      2        1
#> 24:         3        3      2        2
#>     streamset duration stream location