Calculate the Poisson scan statistic.

Calculate the expectation-based Poisson scan statistic by supplying a data.table of observed counts and pre-computed expected value parameters for each location and time. A p-value for the observed scan statistic can be obtained by Monte Carlo simulation.

scan_poisson(table, zones, n_mcsim = 0)

Arguments

table

A data.table with columns location, duration, count, mu. The location column should consist of integers that are unique to each location. The duration column should also consist of integers, starting at 1 for the most recent time period and increasing in reverse chronological order. The column mu should contain the estimated Poisson expected value parameter.

zones

A set of zones, each zone itself a set containing one or more locations of those found in table.

n_mcsim

A non-negative integer; the number of replicate scan statistics to generate in order to calculate a p-value.

Value

An object of class scanstatistics. It has the following fields:

observed

A data.table containing the value of the statistic calculated for each zone-duration combination, for the observed data. The scan statistic is the maximum value of these calculated statistics.

replicated

A numeric vector of length n_mcsim containing the values of the scanstatistics calculated by Monte Carlo simulation.

mlc

A data.table containing the zone, duration, and scanstatistic.

pvalue

The p-value calculated from Monte Carlo replications.

distribution

The assumed distribution of the data; "Poisson" in this case.

type

The type of scan statistic; "Expectation-based" in this case.

zones

The set of zones that was passed to the function as input.

n_locations

The number of locations in the data.

n_zones

The number of zones.

max_duration

The maximum anomaly duration considered.

Details

For the expectation-based Poisson scan statistic, the null hypothesis of no anomaly holds that the count observed at each location \(i\) and duration \(t\) (the number of time periods before present) is Poisson-distributed with expected value \(\mu_{it}\): $$ H_0 : Y_{it} \sim \textrm{Poisson}(\mu_{it}), $$ for all locations \(i = 1, \ldots, m\) and all durations \(t = 1, \ldots,T\), with \(T\) being the maximum duration considered. Under the alternative hypothesis, there is a space-time window \(W\) consisting of a spatial zone \(Z \subset \{1, \ldots, m\}\) and a time window \(D \subseteq \{1, \ldots, T\}\) such that the counts in that window have their expected values inflated by a factor \(q_W > 1\) compared to the null hypothesis: $$ H_1 : Y_{it} \sim \textrm{Poisson}(q_W \mu_{it}), ~~(i,t) \in W. $$ For locations and durations outside of this window, counts are assumed to be distributed as under the null hypothesis. The sets \(Z\) considered are those specified in the argument zones, while the maximum duration \(T\) is taken as the maximum value in the column duration of the input table. For each space-time window \(W\) considered, (the log of) a likelihood ratio is computed using the distributions under the alternative and null hypotheses, and the expectation-based Poisson scan statistic is calculated as the maximum of these quantities over all space-time windows. Point estimates of the parameters \(\mu_{it}\) must be specified in the column mu of the argument table before this function is called.

Examples

# Simple example
set.seed(1)
table <- scanstatistics:::create_table(list(location = 1:4, duration = 1:4),
                                        keys = c("location", "duration"))
table[, mu := 3 * location]

#>     location duration mu
#>  1:        1        1  3
#>  2:        1        2  3
#>  3:        1        3  3
#>  4:        1        4  3
#>  5:        2        1  6
#>  6:        2        2  6
#>  7:        2        3  6
#>  8:        2        4  6
#>  9:        3        1  9
#> 10:        3        2  9
#> 11:        3        3  9
#> 12:        3        4  9
#> 13:        4        1 12
#> 14:        4        2 12
#> 15:        4        3 12
#> 16:        4        4 12
#> 

table[, count := rpois(.N, mu)]

#>     location duration mu count
#>  1:        1        1  3     2
#>  2:        1        2  3     2
#>  3:        1        3  3     3
#>  4:        1        4  3     5
#>  5:        2        1  6     4
#>  6:        2        2  6     9
#>  7:        2        3  6    10
#>  8:        2        4  6     7
#>  9:        3        1  9    10
#> 10:        3        2  9     5
#> 11:        3        3  9     6
#> 12:        3        4  9     6
#> 13:        4        1 12    13
#> 14:        4        2 12    14
#> 15:        4        3 12    13
#> 16:        4        4 12    10
#> 

table[location %in% c(1, 4) & duration < 3, count := rpois(.N, 2 * mu)]

#>     location duration mu count
#>  1:        1        1  3    10
#>  2:        1        2  3     4
#>  3:        1        3  3     3
#>  4:        1        4  3     5
#>  5:        2        1  6     4
#>  6:        2        2  6     9
#>  7:        2        3  6    10
#>  8:        2        4  6     7
#>  9:        3        1  9    10
#> 10:        3        2  9     5
#> 11:        3        3  9     6
#> 12:        3        4  9     6
#> 13:        4        1 12    25
#> 14:        4        2 12    20
#> 15:        4        3 12    13
#> 16:        4        4 12    10
#> 

zones <- scanstatistics:::powerset_zones(4)
result <- scan_poisson(table, zones, 100)
result

#> Data distribution:                Poisson
#> Type of scan statistic:           Expectation-based
#> Number of locations considered:   4
#> Maximum duration considered:      4
#> Number of spatial zones:          15
#> Number of Monte Carlo replicates: 100
#> p-value of observed statistic:    0.01
#> Most likely event duration:       2
#> ID of locations in most likely cluster: 1, 4