Score each location over zones and duration.

For each location, compute the average of the statistic calculated for each space-time window that the location is included in, i.e. average the statistic over both zones and the maximum duration.

score_locations(x)

Arguments

x

An object of class scanstatistic.

Value

A data.table with the following columns:

location

The locations (as integers).

total_score

For each location, the sum of all window statistics that the location appears in.

n_zones

The number of spatial zones that the location appears in.

score

The total score divided by the number of zones and the maximum duration.

relative_score

The score divided by the maximum score.

Examples

# Simple example
set.seed(1)
table <- scanstatistics:::create_table(list(location = 1:4, duration = 1:4),
                                        keys = c("location", "duration"))
table[, mu := 3 * location]

#>     location duration mu
#>  1:        1        1  3
#>  2:        1        2  3
#>  3:        1        3  3
#>  4:        1        4  3
#>  5:        2        1  6
#>  6:        2        2  6
#>  7:        2        3  6
#>  8:        2        4  6
#>  9:        3        1  9
#> 10:        3        2  9
#> 11:        3        3  9
#> 12:        3        4  9
#> 13:        4        1 12
#> 14:        4        2 12
#> 15:        4        3 12
#> 16:        4        4 12
#> 

table[, count := rpois(.N, mu)]

#>     location duration mu count
#>  1:        1        1  3     2
#>  2:        1        2  3     2
#>  3:        1        3  3     3
#>  4:        1        4  3     5
#>  5:        2        1  6     4
#>  6:        2        2  6     9
#>  7:        2        3  6    10
#>  8:        2        4  6     7
#>  9:        3        1  9    10
#> 10:        3        2  9     5
#> 11:        3        3  9     6
#> 12:        3        4  9     6
#> 13:        4        1 12    13
#> 14:        4        2 12    14
#> 15:        4        3 12    13
#> 16:        4        4 12    10
#> 

table[location %in% c(1, 4) & duration < 3, count := rpois(.N, 2 * mu)]

#>     location duration mu count
#>  1:        1        1  3    10
#>  2:        1        2  3     4
#>  3:        1        3  3     3
#>  4:        1        4  3     5
#>  5:        2        1  6     4
#>  6:        2        2  6     9
#>  7:        2        3  6    10
#>  8:        2        4  6     7
#>  9:        3        1  9    10
#> 10:        3        2  9     5
#> 11:        3        3  9     6
#> 12:        3        4  9     6
#> 13:        4        1 12    25
#> 14:        4        2 12    20
#> 15:        4        3 12    13
#> 16:        4        4 12    10
#> 

zones <- scanstatistics:::powerset_zones(4)
result <- scan_poisson(table, zones, 100)
score_locations(result)

#>    location total_score n_zones    score relative_score
#> 1:        1    567.3717       8 17.73036      1.0000000
#> 2:        2    546.1221       8 17.06632      0.9625474
#> 3:        3    513.8509       8 16.05784      0.9056689
#> 4:        4    502.2483       8 15.69526      0.8852192
#>