Get the top (non-overlappig) clusters.

Get the top \(k\) space-time clusters according to the statistic calculated for each cluster (the maximum being the scan statistic). The default is to return the spatially non-overlapping clusters, i.e. those that do not have any locations in common.

top_clusters(x, k = 5, overlapping = FALSE)

Arguments

x

An object of class scanstatistics.

k

An integer, the number of clusters to return

overlapping

Logical; should the top clusters be allowed to overlap in the spatial dimension? The default is FALSE.

Value

A data.table with at most \(k\) rows, with columns zone, duration, statistic.

Examples

set.seed(1)
table <- scanstatistics:::create_table(list(location = 1:4, duration = 1:4),
                                        keys = c("location", "duration"))
table[, mu := 3 * location]

#>     location duration mu
#>  1:        1        1  3
#>  2:        1        2  3
#>  3:        1        3  3
#>  4:        1        4  3
#>  5:        2        1  6
#>  6:        2        2  6
#>  7:        2        3  6
#>  8:        2        4  6
#>  9:        3        1  9
#> 10:        3        2  9
#> 11:        3        3  9
#> 12:        3        4  9
#> 13:        4        1 12
#> 14:        4        2 12
#> 15:        4        3 12
#> 16:        4        4 12

table[, count := rpois(.N, mu)]

#>     location duration mu count
#>  1:        1        1  3     2
#>  2:        1        2  3     2
#>  3:        1        3  3     3
#>  4:        1        4  3     5
#>  5:        2        1  6     4
#>  6:        2        2  6     9
#>  7:        2        3  6    10
#>  8:        2        4  6     7
#>  9:        3        1  9    10
#> 10:        3        2  9     5
#> 11:        3        3  9     6
#> 12:        3        4  9     6
#> 13:        4        1 12    13
#> 14:        4        2 12    14
#> 15:        4        3 12    13
#> 16:        4        4 12    10

table[location %in% c(1, 4) & duration < 3, count := rpois(.N, 2 * mu)]

#>     location duration mu count
#>  1:        1        1  3    10
#>  2:        1        2  3     4
#>  3:        1        3  3     3
#>  4:        1        4  3     5
#>  5:        2        1  6     4
#>  6:        2        2  6     9
#>  7:        2        3  6    10
#>  8:        2        4  6     7
#>  9:        3        1  9    10
#> 10:        3        2  9     5
#> 11:        3        3  9     6
#> 12:        3        4  9     6
#> 13:        4        1 12    25
#> 14:        4        2 12    20
#> 15:        4        3 12    13
#> 16:        4        4 12    10

zones <- scanstatistics:::powerset_zones(4)
result <- scan_poisson(table, zones, 0)
top_clusters(result, k = 4, overlapping = FALSE)

#>    zone duration   statistic
#> 1:    7        2 10.90406367
#> 2:    2        4  0.69430654
#> 3:    3        1  0.05360516