Factor(s) in the product of the EB-ZIP window statistic.

Computes one or more factors in the product of the numerator or denominator of the EB-ZIP window statistic (i.e. the one calculated for a given space-time window W). Should be able to handle cases where p = 0.

zip_statistic_factor(p, d, mu, y, tol = 1e-08)

Arguments

p

Numeric vector of excess zero probabilities.

d

Numeric vector of estimates of the excess zero indicators. Of same length as p.

mu

Numeric vector of given/estimated Poisson expected value parameters. Of same length as p.

y

Integer vector of observed counts. Of same length as p.

tol

Scalar; probability p below this is considered equal to zero.

Value

A numeric vector of same length as input vector p.